Re #1896 minor refactoring and simple unit tests for sort pixels to remember how it works
bin_pixels routine extracted into header to prepare it to be a template (incomplete)
