+++ _LowLevelCode/cpp/test/CMakeLists.txt
@@ -22,6 +22,7 @@ set(TEST_DIRECTORIES
     "combine_sqw.tests"
     "compute_pix_sums.tests"
     "mex_bin_plugin.tests"
+    "sort_pixels_by_bins.tests"
 )
 foreach(_test_dir ${TEST_DIRECTORIES})
     add_subdirectory("${_test_dir}")
